Update bm.bat
Fixed an issue where the url append file was being unconditionally called, regardless of whether or not existed if the url flag was set. Added a check to see if the append file exists to prevent an error from being printed to console. Had little to no real impact on the script working with URLs, but did print an error to console about the system not being able to find the file specified.