Update bm.bat
Doing a hard deletion on the append file before checking for a %~2 parameter. There was an unintended behavior of loading the append url file regardless of which bookmark was opened if the url flag was defined, as long as it still existed. Now the script will delete that file if it exists before running the second parameter check.